pyproject.toml
[tool.poetry]
name = "medsignal"
version = "0.1.0"
description = "Drug Safety Signal Detection Platform"
authors = ["pradhanprac"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = ">=0.135.3,<0.136.0"
uvicorn = ">=0.44.0,<0.45.0"
requests = ">=2.33.1,<3.0.0"
kafka-python = ">=2.3.1,<3.0.0"
psycopg2-binary = ">=2.9.11,<3.0.0"
python-dotenv = ">=1.0.0,<2.0.0"
pandas = ">=2.0.0,<3.0.0"
sqlalchemy = ">=2.0.49,<3.0.0"
biopython = ">=1.87,<2.0"
sentence-transformers = ">=2.7.0,<3.0.0"
chromadb = ">=0.5.0,<2.0.0"
structlog = ">=25.5.0,<26.0.0"
openai = ">=1.0.0,<2.0.0"
langgraph = ">=0.2.0,<1.0.0"
pydantic = ">=2.0.0,<3.0.0"
pytest = ">=8.0.0,<9.0.0"
redis = "^7.4.0"
snowflake-connector-python = "^4.4.0"
rank-bm25 = "^0.2.2"
prometheus-client = "^0.25.0"
plotly = "^6.7.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
markers = [
    "unit: pure logic tests, no external dependencies",
    "live: integration tests, requires ChromaDB loaded",
]